+++ rolling/arithmetic/nunique.py
@@ -42,13 +42,13 @@ class Nunique(RollingObject):
 
     def _init_fixed(self, iterable, window_size, **kwargs):
         head = islice(self._iterator, window_size - 1)
-        self._buffer = deque(head, maxlen=window_size)
+        self._buffer = deque(head)
         # append a dummy value that is removed when next() is called
         self._buffer.appendleft("dummy_value")
         self._counter = Counter(self._buffer)
 
     def _init_variable(self, iterable, window_size, **kwargs):
-        self._buffer = deque(maxlen=window_size)
+        self._buffer = deque()
         self._counter = Counter()
 
     def _update_window(self, new):